F450 tweaks, Joystick SDL2
F450 landing gear is now a tripod. JSBSim was having trouble with the vehicle on the ground, reporting 0g and -2g verticle acceleration which was messing up the EKF.

Dropped pygames in favor of SDL2 for joystick interface. pygames was coming up with a strange set of axes on Windows.